refactor CountryIpModal
