+++ src/features/register/CountryIpModal/index.tsx
@@ -1,0 +1,2 @@
+import CountryIpModal from './CountryIpModal';
+export default CountryIpModal;
